+++ src/components/cell-control/autocompleter.jsx
@@ -1,15 +1,16 @@
 import PropTypes from 'prop-types';
-import React, { useRef } from 'react';
+import React, { forwardRef, useRef } from 'react';
 import CellControl from './cell-control.jsx';
 import AutocompleterControl from '../control/autocompleter.jsx';
 
-function Autocompleter(props) {
+const Autocompleter = forwardRef(function Autocompleter(props, ref) {
   const refs = {
     container: useRef(),
   };
 
   return (
     <CellControl
+      className={props.className}
       labelledBy={props.labelledBy}
       readOnly={props.readOnly}
       ref={refs.container}
@@ -24,7 +25,7 @@ function Autocompleter(props) {
         onChange={props.onChange}
         placeholder={props.placeholder}
         readOnly={props.readOnly}
-        ref={refs.input}
+        ref={ref}
         required={props.required}
         searchParam={props.searchParam}
         validationMessage={props.validationMessage}
@@ -33,11 +34,13 @@ function Autocompleter(props) {
       />
     </CellControl>
   );
-}
+});
 
 Autocompleter.propTypes = {
   /** `apiEndpoint` should be the route of the api's endpoint (excluding the base api), eg. `/workspaces`. */
   apiEndpoint: PropTypes.string,
+  /** A class name for the table cell container. */
+  className: PropTypes.string,
   /** displayValueEvaluator is handled if the key following: `title`, `name`, `full_name`, `currency`; Otherwise, pass in something like `displayValueEvaluator: (model) -> { model.rate_card_name }` */
   displayValueEvaluator: PropTypes.func,
   id: PropTypes.string.isRequired,