+++ src/components/cell-control/input.jsx
@@ -1,11 +1,12 @@
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, { forwardRef } from 'react';
 import CellControl from '../cell-control/cell-control.jsx';
 import InputControl from '../control/input.jsx';
 
-function Input(props) {
+const Input = forwardRef(function Input(props, ref) {
   return (
     <CellControl
+      className={props.className}
       labelledBy={props.labelledBy}
       readOnly={props.readOnly}
     >
@@ -13,14 +14,17 @@ function Input(props) {
         id={props.id}
         labelledBy={props.labelledBy}
         readOnly={props.readOnly}
+        ref={ref}
         required={props.required}
         validationMessage={props.validationMessage}
       />
     </CellControl>
   );
-}
+});
 
 Input.propTypes = {
+  /** A class name for the table cell container. */
+  className: PropTypes.string,
   /** A unique ID for the component. */
   id: PropTypes.string.isRequired,
   /** The ID of the column header. */
@@ -34,6 +38,7 @@ Input.propTypes = {
 };
 
 Input.defaultProps = {
+  className: undefined,
   readOnly: false,
   required: false,
   validationMessage: '',